Changes applied:
//If with this variable what you wanted is to declare a constant, the correct way to do it is:

public static final STUDENTNAME = "Benjamin";

But, are you sure you want to have a report card just for one student?

Wouldn't it be better to declare this variable as the rest of them with its getter/setter methods to use with all students? :smile:

//Please, modify it, instantiate it in the constructor and add setter/getter methods to it.

You must override the toStringmethod from its superclass.
Please, take a look at this good article about this.